requirements.txt
discord.py
mock
nose
pygithub
python-Levenshtein
requests
tinydb
youtube-dl
